requirements.txt
kafka-python==2.0.2
pluggy==1.4.0
pytest==8.1.1
setuptools==58.1.0
tomli==2.0.1
pytest==8.1.1
requests==2.31.0
black==24.3.0